Rewrote the command inference to be less klugey.
